+++ dotnet/src/Agents/AzureAI/Extensions/KernelFunctionExtensions.cs
@@ -1,7 +1,5 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using Azure.AI.Projects;
 
 namespace Microsoft.SemanticKernel.Agents.AzureAI;
@@ -16,78 +14,13 @@ internal static class KernelFunctionExtensions
     /// <returns>An OpenAI tool definition</returns>
     public static FunctionToolDefinition ToToolDefinition(this KernelFunction function, string pluginName)
     {
-        var metadata = function.Metadata;
-        if (metadata.Parameters.Count > 0)
+        if (function.Metadata.Parameters.Count > 0)
         {
-            var required = new List<string>(metadata.Parameters.Count);
-            var parameters =
-                metadata.Parameters.ToDictionary(
-                    p => p.Name,
-                    p =>
-                    {
-                        if (p.IsRequired)
-                        {
-                            required.Add(p.Name);
-                        }
+            BinaryData parameterData = function.Metadata.CreateParameterSpec();
 
-                        return
-                            new
-                            {
-                                type = ConvertType(p.ParameterType),
-                                description = p.Description,
-                            };
-                    });
-
-            var spec =
-                new
-                {
-                    type = "object",
-                    properties = parameters,
-                    required,
-                };
-
-            return new FunctionToolDefinition(FunctionName.ToFullyQualifiedName(function.Name, pluginName), function.Description, BinaryData.FromObjectAsJson(spec));
+            return new FunctionToolDefinition(FunctionName.ToFullyQualifiedName(function.Name, pluginName), function.Description, parameterData);
         }
 
         return new FunctionToolDefinition(FunctionName.ToFullyQualifiedName(function.Name, pluginName), function.Description);
-    }
-
-    private static string ConvertType(Type? type)
-    {
-        if (type is null || type == typeof(string))
-        {
-            return "string";
-        }
-
-        if (type == typeof(bool))
-        {
-            return "boolean";
-        }
-
-        if (type.IsEnum)
-        {
-            return "enum";
-        }
-
-        if (type.IsArray)
-        {
-            return "array";
-        }
-
-        if (type == typeof(DateTime) || type == typeof(DateTimeOffset))
-        {
-            return "date-time";
-        }
-
-        return Type.GetTypeCode(type) switch
-        {
-            TypeCode.SByte or TypeCode.Byte or
-            TypeCode.Int16 or TypeCode.UInt16 or
-            TypeCode.Int32 or TypeCode.UInt32 or
-            TypeCode.Int64 or TypeCode.UInt64 or
-            TypeCode.Single or TypeCode.Double or TypeCode.Decimal => "number",
-
-            _ => "object",
-        };
     }
 }
